Move DronePlayer to a seperate file
